BUG#28261178 [4/5] cosmetic: Moved ACL-checking test helper function to a common place to allow reuse
RB#21559
