+++ router/tests/CMakeLists.txt
@@ -1,4 +1,4 @@
-# Copyright (c) 2015, 2018, Oracle and/or its affiliates. All rights reserved.
+# Copyright (c) 2015, 2019, Oracle and/or its affiliates. All rights reserved.
 #
 # This program is free software; you can redistribute it and/or modify
 # it under the terms of the GNU General Public License, version 2.0,
@@ -29,6 +29,7 @@ INCLUDE_DIRECTORIES(helpers
   )
 
 SET(helper_sources
+  helpers/filesystem_utils.cc
   helpers/gtest_testname.cc
   helpers/mysql_session_replayer.cc
   helpers/router_component_test.cc
